% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\alias{flush_cache}
\title{Flush \code{basemaps} cache}
\usage{
flush_cache()
}
\value{
None.
}
\description{
This function flushes the \code{basemaps} cache and thereby removes all previously queried and/or composited products from the map directories (temporary or user-defined using the argument \code{map_dir}) used during the current session.
}
\examples{
library(basemaps)
flush_cache()

}
