% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{ext}
\alias{ext_eur}
\title{Example extent}
\format{
\code{sf} object

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
data(ext)

data(ext_eur)
}
\description{
The example datasets contain the \code{sf} objects \code{ext} and \code{ext_eur} that can be used to call \code{\link{basemap}} and the associated functions.
}
\keyword{datasets}
