% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{draw_ext}
\alias{draw_ext}
\title{Draw extent}
\usage{
draw_ext()
}
\value{
An \code{sf} object
}
\description{
This function lets you draw an extent on an interactive map. It is a simple wrapper around \code{mapedit::drawFeatures()} written by Tim Appelhans et al.
}
\examples{
\dontrun{
library(basemaps)

# draw extent interactively
ext <- draw_ext()

# set defaults for the basemap
set_defaults(ext = ext, map_service = "osm", map_type = "terrain_bg")
# for mapbox maps, you need a map_token. Register for free at mapbox.com to get a token

# load and return basemap map as raster (default)
map <- basemap()
}

}
