% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ext}
\alias{ext}
\title{Example extemt}
\format{
\code{sf} object
}
\usage{
data(ext)
}
\description{
The example dataset contains the object \code{ext}, an \code{sf} object that can be used to call \code{\link{basemap}} and the associated functions.
}
\keyword{datasets}
