% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{basemap}
\alias{basemap}
\title{Get a spatial basemap}
\usage{
basemap(
  ext,
  map_service = NA,
  map_type = NA,
  map_res = NA,
  map_token = NA,
  map_dir = NA,
  class = "raster",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{ext}{extent to be covered by the basemap. Any spatial class.}

\item{map_service}{character,}

\item{map_type}{character,}

\item{map_res}{numeric,}

\item{map_token}{character,}

\item{map_dir}{character, cache directory}

\item{class}{character, output class.}

\item{...}{additional arguments.}

\item{verbose}{}
}
\value{

}
\description{
This function (down)loads and caches a basemap from a defined extent \code{ext}, \code{map_service} and \code{map_type} and returns it as a object of a defined \code{class}.
}
