% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{set_defaults}
\alias{get_defaults}
\title{Set \code{basemaps} defaults}
\usage{
set_defaults(
  map_service = NULL,
  map_type = NULL,
  map_res = NULL,
  map_token = NULL,
  map_dir = NULL
)

get_defaults()
}
\arguments{
\item{map_service}{character,}

\item{map_type}{character,}

\item{map_res}{numeric,}

\item{map_token}{character,}

\item{map_dir}{character, cache directory}
}
\value{
None
}
\description{
This function sets the defaults of all map arguments passed to \code{\link{basemap}} and associated functions.
}
